\name{GSE280465-package}
\alias{GSE280465-package}
\docType{package}
\title{GSE280465 EPICv2 methylation data in ExperimentHub}
\description{
The \pkg{GSE280465} package provides ExperimentHub access to a processed
adult DNA methylation resource derived from GEO accession GSE280465.
}
\details{
The hosted resource is a \code{RangedSummarizedExperiment} created from an
Illumina Infinium MethylationEPIC version 2 \code{GenomicRatioSet}. It contains
beta, M, and CN assays for 163 adult samples from 47 individuals across buccal,
saliva, dried blood spot, and peripheral blood mononuclear cell tissues.

The resource is hosted on Zenodo at
\url{https://zenodo.org/records/21200586} and is registered for access through
\pkg{ExperimentHub}.

See \code{browseVignettes("GSE280465")} for a usage example.
}
\seealso{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE280465}
}
\keyword{datasets}
